Spring validation have implemented

+++ src/main/java/com/alexbro/onlinebank/facade/account/AccountFacade.java
@@ -3,13 +3,12 @@ package com.alexbro.onlinebank.facade.account;
 import com.alexbro.onlinebank.facade.data.account.AccountData;
 import com.alexbro.onlinebank.facade.data.exchange.ExchangeData;
 
-import java.math.BigDecimal;
 import java.util.List;
 import java.util.Optional;
 
 public interface AccountFacade {
 
-    void transfer(String accountCode, Long cardNumber, BigDecimal sum);
+    void transfer(String accountCode, Long cardNumber, Double sum);
 
     List<AccountData> findAllByCurrency(String currencyCode);
 
@@ -17,8 +16,8 @@ public interface AccountFacade {
 
     Optional<AccountData> findByCardNumber(Long cardNumber);
 
-    ExchangeData getExchangeData(AccountData accountFrom, AccountData accountTo, BigDecimal sum);
+    ExchangeData getExchangeData(AccountData accountFrom, AccountData accountTo, Double sum);
 
-    void exchange(String accountFromCode, String accountToCode, BigDecimal sum);
+    void exchange(String accountFromCode, String accountToCode, Double sum);
 
 }
